[refactor] DDD 구조 변경 #63
