+++ src/main/java/classfit/example/classfit/attendance/controller/AttendanceController.java
@@ -4,9 +4,9 @@ import classfit.example.classfit.attendance.dto.request.StudentAttendanceUpdateR
 import classfit.example.classfit.attendance.dto.response.StudentAttendanceResponse;
 import classfit.example.classfit.attendance.service.AttendanceService;
 import classfit.example.classfit.attendance.service.AttendanceUpdateService;
+import classfit.example.classfit.classStudent.domain.ClassStudent;
 import classfit.example.classfit.common.ApiResponse;
-import classfit.example.classfit.domain.ClassStudent;
-import classfit.example.classfit.domain.Student;
+import classfit.example.classfit.student.domain.Student;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -28,10 +28,10 @@ public class AttendanceController {
     @GetMapping("/")
     @Operation(summary = "전체 출결 관리 조회", description = "전체 학생을 클릭 시 조회되는 출결 api 입니다.")
     public ApiResponse<List<StudentAttendanceResponse>> getAttendance(
-            @Parameter(description = "이전 또는 이후 출결 조회 시 필요한 값으로, 현재는 0(디폴트), 이전 주는 음수로, 다음 주는 양수로 표시합니다. 4주 전부터 2주 후까지(-4 ~ +2) 조회 가능합니다.")
-            @RequestParam(value = "weekOffset", defaultValue = "0") int weekOffset,
-            @Parameter(description = "조회할 페이지 번호로, 기본값은 0이며, 페이지는 0부터 시작합니다.")
-            @RequestParam(value = "page", defaultValue = "0") int page) {
+        @Parameter(description = "이전 또는 이후 출결 조회 시 필요한 값으로, 현재는 0(디폴트), 이전 주는 음수로, 다음 주는 양수로 표시합니다. 4주 전부터 2주 후까지(-4 ~ +2) 조회 가능합니다.")
+        @RequestParam(value = "weekOffset", defaultValue = "0") int weekOffset,
+        @Parameter(description = "조회할 페이지 번호로, 기본값은 0이며, 페이지는 0부터 시작합니다.")
+        @RequestParam(value = "page", defaultValue = "0") int page) {
         List<LocalDate> weekRange = attendanceService.getWeeklyAttendanceRange(weekOffset);
         Page<Student> students = attendanceService.getAllStudents(page);
         List<StudentAttendanceResponse> studentAttendances = attendanceService.getStudentAttendance(students.getContent(), weekRange);
@@ -41,12 +41,12 @@ public class AttendanceController {
     @GetMapping("/{mainClassId}/{subClassId}")
     @Operation(summary = "특정 클래스 출결 관리 조회", description = "특정 클래스를 클릭 시 조회되는 출결 api 입니다.")
     public ApiResponse<List<StudentAttendanceResponse>> getClassAttendance(
-            @Parameter(description = "이전 또는 이후 출결 조회 시 필요한 값으로, 현재는 0(디폴트), 이전 주는 음수로, 다음 주는 양수로 표시합니다. 4주 전부터 2주 후까지(-4 ~ +2) 조회 가능합니다.")
-            @RequestParam(value = "weekOffset", defaultValue = "0") int weekOffset,
-            @Parameter(description = "조회할 페이지 번호로, 기본값은 0이며, 페이지는 0부터 시작합니다.")
-            @RequestParam(value = "page", defaultValue = "0") int page,
-            @PathVariable("mainClassId") Long mainClassId,
-            @PathVariable("subClassId") Long subClassId) {
+        @Parameter(description = "이전 또는 이후 출결 조회 시 필요한 값으로, 현재는 0(디폴트), 이전 주는 음수로, 다음 주는 양수로 표시합니다. 4주 전부터 2주 후까지(-4 ~ +2) 조회 가능합니다.")
+        @RequestParam(value = "weekOffset", defaultValue = "0") int weekOffset,
+        @Parameter(description = "조회할 페이지 번호로, 기본값은 0이며, 페이지는 0부터 시작합니다.")
+        @RequestParam(value = "page", defaultValue = "0") int page,
+        @PathVariable("mainClassId") Long mainClassId,
+        @PathVariable("subClassId") Long subClassId) {
         List<LocalDate> weekRange = attendanceService.getWeeklyAttendanceRange(weekOffset);
         Page<ClassStudent> students = attendanceService.getClassStudentsByMainClassAndSubClass(mainClassId, subClassId, page);
         List<StudentAttendanceResponse> studentAttendances = attendanceService.getStudentAttendance(students.getContent(), weekRange);
