+++ src/main/java/classfit/example/classfit/attendance/domain/AttendanceStatus.java
@@ -1,4 +1,4 @@
-package classfit.example.classfit.common;
+package classfit.example.classfit.attendance.domain;
 
 public enum AttendanceStatus {
     PRESENT("출석"), ABSENT("결석"), LATE("지각");
