+++ src/main/java/classfit/example/classfit/category/domain/SubClass.java
@@ -1,26 +1,21 @@
-package classfit.example.classfit.domain;
+package classfit.example.classfit.category.domain;
 
-import classfit.example.classfit.common.BaseEntity;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
+import classfit.example.classfit.classStudent.domain.ClassStudent;
+import classfit.example.classfit.common.domain.BaseEntity;
+import classfit.example.classfit.member.domain.Member;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class SubClass extends BaseEntity {
+
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -41,14 +36,14 @@ public class SubClass extends BaseEntity {
     @OneToMany(mappedBy = "subClass", cascade = CascadeType.ALL, orphanRemoval = true)
     private final List<ClassStudent> classStudents = new ArrayList<>();
 
-    // 업데이트 관련 메서드
-    public void updateSubClassName(String subClassName) {
-        this.subClassName = subClassName;
-    }
-
-    public SubClass(String subClassName, Member member,MainClass mainClass) {
+    public SubClass(String subClassName, Member member, MainClass mainClass) {
         this.subClassName = subClassName;
         this.member = member;
         this.mainClass = mainClass;
+    }
+
+    // 업데이트 관련 메서드
+    public void updateSubClassName(String subClassName) {
+        this.subClassName = subClassName;
     }
 }
