+++ src/main/java/classfit/example/classfit/category/repository/MainClassRespository.java
@@ -1,14 +1,14 @@
 package classfit.example.classfit.category.repository;
 
-import classfit.example.classfit.domain.MainClass;
-import classfit.example.classfit.domain.Member;
+import classfit.example.classfit.category.domain.MainClass;
+import classfit.example.classfit.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface MainClassRespository extends JpaRepository<MainClass,Long> {
+public interface MainClassRespository extends JpaRepository<MainClass, Long> {
     List<MainClass> findAllByOrderByMainClassNameAsc();
 
     boolean existsByMemberAndMainClassName(Member member, String mainClassName);
