+++ src/main/java/classfit/example/classfit/category/service/ClassInfoService.java
@@ -1,10 +1,10 @@
 package classfit.example.classfit.category.service;
 
+import classfit.example.classfit.category.domain.MainClass;
+import classfit.example.classfit.category.domain.SubClass;
 import classfit.example.classfit.category.dto.response.ClassInfoResponse;
 import classfit.example.classfit.category.dto.response.SubClassResponse;
 import classfit.example.classfit.category.repository.MainClassRespository;
-import classfit.example.classfit.domain.MainClass;
-import classfit.example.classfit.domain.SubClass;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,15 +21,15 @@ public class ClassInfoService {
     public List<ClassInfoResponse> getClasses() {
         List<MainClass> mainClasses = mainClassRespository.findAllByOrderByMainClassNameAsc();
         return mainClasses.stream()
-                .map(this::mapToClassInfoResponse)
-                .collect(Collectors.toList());
+            .map(this::mapToClassInfoResponse)
+            .collect(Collectors.toList());
     }
 
     private ClassInfoResponse mapToClassInfoResponse(MainClass mainClass) {
         List<SubClassResponse> subClassResponses = mainClass.getSubClasses().stream()
-                .sorted((subClass1, subClass2) -> subClass1.getSubClassName().compareTo(subClass2.getSubClassName()))
-                .map(subClass -> mapToSubClassResponse(mainClass.getId(), subClass))
-                .collect(Collectors.toList());
+            .sorted((subClass1, subClass2) -> subClass1.getSubClassName().compareTo(subClass2.getSubClassName()))
+            .map(subClass -> mapToSubClassResponse(mainClass.getId(), subClass))
+            .collect(Collectors.toList());
         return new ClassInfoResponse(mainClass.getId(), mainClass.getMainClassName(), subClassResponses);
     }
