+++ src/main/java/classfit/example/classfit/classStudent/domain/ClassStudent.java
@@ -1,14 +1,8 @@
-package classfit.example.classfit.domain;
+package classfit.example.classfit.classStudent.domain;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
+import classfit.example.classfit.category.domain.SubClass;
+import classfit.example.classfit.student.domain.Student;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
