--- src/main/java/classfit/example/classfit/common/Gender.java
@@ -1,5 +1,0 @@
-package classfit.example.classfit.common;
-
-public enum Gender {
-    MALE, FEMALE, NONE
-}
