+++ src/main/java/classfit/example/classfit/common/domain/ResultType.java
@@ -1,4 +1,4 @@
-package classfit.example.classfit.common;
+package classfit.example.classfit.common.domain;
 
 public enum ResultType {
     SUCCESS,
