--- src/main/java/classfit/example/classfit/member/MemberRepository.java
@@ -1,10 +1,0 @@
-package classfit.example.classfit.member;
-
-import classfit.example.classfit.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member,Long> {
-
-}
