+++ src/main/java/classfit/example/classfit/member/domain/Gender.java
@@ -1,0 +1,5 @@
+package classfit.example.classfit.member.domain;
+
+public enum Gender {
+    MALE, FEMALE, NONE
+}
