+++ src/main/java/classfit/example/classfit/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package classfit.example.classfit.member.repository;
+
+import classfit.example.classfit.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+}
