+++ src/main/java/classfit/example/classfit/student/repository/StudentRepository.java
@@ -1,14 +1,15 @@
 package classfit.example.classfit.student.repository;
 
-import classfit.example.classfit.domain.Student;
-import java.util.List;
-import java.util.Optional;
+import classfit.example.classfit.student.domain.Student;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface StudentRepository extends JpaRepository<Student, Long> {
